Tracked Commit #4
- Fix resistance text icon
- Optimize all tiles that output stuff into tiles by storing the tiles instead of finding em every time.
- Improve damage taken by wires algorithm
- Fix wires so that instead of having a constant current they only have a constant energy flow. Aka joules. This fixes it so that amps actually change alongside voltage when transformed.
- Add power transfer in the multimeter.
- Add obsidian dust item. Crafted in the obsidian crusher.
- Finally add ultimate circuit recipe.
- Add crash that caused all machines to crash on usage
- Cleanup code

+++ src/main/java/electrodynamics/common/tile/generic/GenericTileProcessor.java
@@ -96,7 +96,7 @@ public abstract class GenericTileProcessor extends GenericTileInventory implemen
 			case 2:
 				return (int) Math.ceil(getJoulesPerTick());
 			case 3:
-				return (int) getRequiredTicks() == 0 ? 1 : getRequiredTicks();
+				return getRequiredTicks() == 0 ? 1 : getRequiredTicks();
 			default:
 				return 0;
 			}
